+++ ios/Classes/PlatformInterface/VideoInfo.swift
@@ -1,7 +1,7 @@
 struct VideoInfo {
     let height: Int
     let width: Int
-    let duration: Int
+    let duration: Int64
 
     func toMap() -> [String: Any] {
         [
